[le_2d] refactor: hide encode_to_bytes - `encode_to_bytes` is an internal function only used by the 2d   rasterizer, we don't want it publicly accessible. - moved encode_to_bytes into le_2d.cpp - extended le_2d_shared.h to contain all necessary internal definitions
