+++ modules/le_2d/le_2d_encoder.cpp
@@ -25,15 +25,10 @@ Vello shaders are licensed under Apache License, Version 2.0, MIT, or Unilicense
 
  */
 
-#include "le_2d.h"
 #include "le_log.h"
-
-#include <cassert>
-#include <cmath>
-#include <vector>
-// ----------------------------------------------------------------------
 #include "private/le_2d/le_2d_shared.h"
-#include "glm/vec2.hpp"
+
+// ----------------------------------------------------------------------
 
 static constexpr auto FRAC_PI_2 = 1.57079632679489661923132169163975144; // 1.57079637f32
 static constexpr auto TWO_PI    = 6.28318530717958647692528676655900576;
@@ -44,55 +39,6 @@ static auto logger() {
 }
 
 // ----------------------------------------------------------------------
-
-// enum class PathSegmentType : uint8_t {
-//	eUndefined = 0,
-//	eLineTo    = 1,
-//	eQuadTo    = 2,
-//	eCubicTo   = 3,
-// };
-
-enum class PathTag : uint8_t {
-	/// Bit for path segments that are represented as f32 values. If unset
-	/// they are represented as i16.
-	F32_BIT = 0x8,
-	/// Mask for bottom 3 bits that contain the [`PathSegmentType`].
-	SEGMENT_MASK = 0x3,
-
-	/// 16-bit integral line segment.
-	LINE_TO_I16 = uint8_t( 1 ),
-
-	/// 16-bit integral quadratic segment.
-	QUAD_TO_I16 = uint8_t( 2 ),
-
-	/// 16-bit integral cubic segment.
-	CUBIC_TO_I16 = uint8_t( 3 ),
-
-	/// 32-bit floating point line segment.
-	LINE_TO_F32 = LINE_TO_I16 | F32_BIT,
-
-	/// 32-bit floating point quadratic segment.
-	QUAD_TO_F32 = QUAD_TO_I16 | F32_BIT,
-
-	/// 32-bit floating point cubic segment.
-	CUBIC_TO_F32 = CUBIC_TO_I16 | F32_BIT,
-
-	/// Transform marker.
-	TRANSFORM = 0x20,
-
-	/// Path marker.
-	PATH = 0x10,
-
-	/// Style setting.
-	STYLE = 0x40,
-
-	/// Bit that marks a segment that is the end of a subpath.
-	SUBPATH_END_BIT = 0x4,
-};
-
-// static inline PathSegmentType get_path_segment_type( PathTag const& e ) {
-// 	return PathSegmentType( uint8_t( e ) & uint8_t( PathTag::SEGMENT_MASK ) );
-// }
 
 static inline bool is_path_segment( PathTag const& e ) {
 	return uint8_t( e ) & uint8_t( PathTag::SEGMENT_MASK );
@@ -111,333 +57,7 @@ static inline bool is_f32( PathTag const& e ) {
 }
 
 using namespace le_2d;
-
-struct Style {
-
-	// ***********************************************************************
-	//
-	// The float 32->float 16 conversion function is by Fabian "ryg" Giesen.
-	// it was released under public domain and copied from here:
-	//
-	// https://gist.github.com/rygorous/2156668#file-gistfile1-cpp-L285
-	//
-
-	union FP32 {
-		uint32_t u;
-		float    f;
-		struct
-		{
-			uint32_t Mantissa : 23;
-			uint32_t Exponent : 8;
-			uint32_t Sign : 1;
-		};
-	};
-
-	union FP16 {
-		unsigned short u;
-		struct
-		{
-			uint32_t Mantissa : 10;
-			uint32_t Exponent : 5;
-			uint32_t Sign : 1;
-		};
-	};
-
-	static FP16 float_to_half_fast3( FP32 f ) {
-		FP32     f32infty   = { 255 << 23 };
-		FP32     f16infty   = { 31 << 23 };
-		FP32     magic      = { 15 << 23 };
-		uint32_t sign_mask  = 0x80000000u;
-		uint32_t round_mask = ~0xfffu;
-		FP16     o          = { 0 };
-
-		uint32_t sign = f.u & sign_mask;
-		f.u ^= sign;
-
-		// NOTE all the integer compares in this function can be safely
-		// compiled into signed compares since all operands are below
-		// 0x80000000. Important if you want fast straight SSE2 code
-		// (since there's no unsigned PCMPGTD).
-
-		if ( f.u >= f32infty.u )                          // Inf or NaN (all exponent bits set)
-			o.u = ( f.u > f32infty.u ) ? 0x7e00 : 0x7c00; // NaN->qNaN and Inf->Inf
-		else                                              // (De)normalized number or zero
-		{
-			f.u &= round_mask;
-			f.f *= magic.f;
-			f.u -= round_mask;
-			if ( f.u > f16infty.u )
-				f.u = f16infty.u; // Clamp to signed infinity if overflowed
-
-			o.u = f.u >> 13; // Take the bits!
-		}
-
-		o.u |= sign >> 16;
-		return o;
-	}
-	// ***********************************************************************
-
-	///
-	/// This is taken/adapted from vello_encoding, which can be found here:
-	/// <https://github.com/linebender/vello/tree/main/vello_encoding>
-	///
-	/// Encodes the stroke and fill style parameters. This field is split into two 16-bit
-	/// parts:
-	///
-	/// - `flags: u16` - encodes fill vs stroke, even-odd vs non-zero fill
-	///   mode for fills and cap and join style for strokes. See the
-	///   `FLAGS_*` constants below for more information.
-	///
-	///   ```text
-	///   flags: |style|fill|join|start cap|end cap|reserved|
-	///    bits:  0     1    2-3  4-5       6-7     8-15
-	///   ```
-	///
-	/// - `miter_limit: u16` - The miter limit for a stroke, encoded in
-	///   binary16 (half) floating point representation. This field is
-	///   only meaningful for the `Join::Miter` join style. It's ignored
-	///   for other stroke styles and fills.
-	uint32_t flags_and_miter_limit;
-	//
-	// 	                               /// Encodes the stroke width. This field is ignored for fills.
-	float line_width;
-
-	/// 0 for a fill, 1 for a stroke
-	static constexpr uint32_t FLAGS_STYLE_BIT = 0x8000'0000;
-
-	/// 0 for non-zero, 1 for even-odd
-	static constexpr uint32_t FLAGS_FILL_BIT = 0x4000'0000;
-
-	/// Encodings for join style:
-	///    - 0b00 -> bevel
-	///    - 0b01 -> miter
-	///    - 0b10 -> round
-	static constexpr uint32_t FLAGS_JOIN_BITS_BEVEL = 0;
-	static constexpr uint32_t FLAGS_JOIN_BITS_MITER = 0x1000'0000;
-	static constexpr uint32_t FLAGS_JOIN_BITS_ROUND = 0x2000'0000;
-	static constexpr uint32_t FLAGS_JOIN_MASK       = 0x3000'0000;
-
-	/// Encodings for cap style:
-	///    - 0b00 -> butt
-	///    - 0b01 -> square
-	///    - 0b10 -> round
-	static constexpr uint32_t FLAGS_CAP_BITS_BUTT   = 0;
-	static constexpr uint32_t FLAGS_CAP_BITS_SQUARE = 0x0100'0000;
-	static constexpr uint32_t FLAGS_CAP_BITS_ROUND  = 0x0200'0000;
-
-	static constexpr uint32_t FLAGS_START_CAP_BITS_BUTT   = FLAGS_CAP_BITS_BUTT << 2;
-	static constexpr uint32_t FLAGS_START_CAP_BITS_SQUARE = FLAGS_CAP_BITS_SQUARE << 2;
-	static constexpr uint32_t FLAGS_START_CAP_BITS_ROUND  = FLAGS_CAP_BITS_ROUND << 2;
-	static constexpr uint32_t FLAGS_END_CAP_BITS_BUTT     = FLAGS_CAP_BITS_BUTT;
-	static constexpr uint32_t FLAGS_END_CAP_BITS_SQUARE   = FLAGS_CAP_BITS_SQUARE;
-	static constexpr uint32_t FLAGS_END_CAP_BITS_ROUND    = FLAGS_CAP_BITS_ROUND;
-
-	static constexpr uint32_t FLAGS_START_CAP_MASK = 0x0C00'0000;
-	static constexpr uint32_t FLAGS_END_CAP_MASK   = 0x0300'0000;
-	static constexpr uint32_t MITER_LIMIT_MASK     = 0xFFFF;
-
-	static Style from_fill( FillStyle const& fill ) {
-		return Style{
-		    .flags_and_miter_limit = ( fill == FillStyle::EvenOdd ) ? FLAGS_FILL_BIT : 0,
-		    .line_width            = 0.f,
-		};
-	};
-
-	static Style from_stroke( Stroke const& stroke ) {
-		uint32_t style     = FLAGS_STYLE_BIT;
-		uint32_t join      = 0;
-		uint32_t start_cap = 0;
-		uint32_t end_cap   = 0;
-		// clang-format off
-		switch ( stroke.join ) {
-			case Join::eBevel: join = FLAGS_JOIN_BITS_BEVEL; break;
-			case Join::eMiter: join = FLAGS_JOIN_BITS_MITER; break;
-			case Join::eRound: join = FLAGS_JOIN_BITS_ROUND; break;
-		};
-		switch ( stroke.start_cap ) {
-			case Cap::eButt:   start_cap = FLAGS_START_CAP_BITS_BUTT; break;
-			case Cap::eRound:  start_cap = FLAGS_START_CAP_BITS_ROUND; break;
-			case Cap::eSquare: start_cap = FLAGS_START_CAP_BITS_SQUARE; break;
-		};
-		switch ( stroke.end_cap) {
-			case Cap::eButt:   end_cap = FLAGS_END_CAP_BITS_BUTT; break;
-			case Cap::eRound:  end_cap = FLAGS_END_CAP_BITS_ROUND; break;
-			case Cap::eSquare: end_cap = FLAGS_END_CAP_BITS_SQUARE; break;
-		};
-		// clang-format on
-		uint32_t miter_limit = float_to_half_fast3( FP32( stroke.miter_limit ) ).u;
-
-		return Style{
-		    .flags_and_miter_limit = style | join | start_cap | end_cap | miter_limit,
-		    .line_width            = float( stroke.width ),
-		};
-	}
-
-	const bool operator==( Style const& rhs ) const {
-		return ( 0 == memcmp( this, &rhs, sizeof( Style ) ) );
-	}
-
-	const bool operator!=( Style const& rhs ) const {
-		return !( *this == rhs );
-	}
-};
-
-enum class DrawTag : uint32_t {
-	NOP             = 0,
-	COLOUR          = 0x44,
-	LINEAR_GRADIENT = 0x114,
-	RADIAL_GRADIENT = 0x29c,
-	SWEEP_GRADIENT  = 0x254,
-	IMAGE           = 0x28c,
-	BLUR_RECT       = 0x2d4,
-	BEGIN_CLIP      = 0x9,
-	END_CLIP        = 0x21,
-};
-
-static uint32_t draw_tag_get_info_size( DrawTag const& t ) {
-	return ( ( uint32_t( t ) >> 6 ) & uint32_t( 0xf ) );
-}
-
-enum ExtendMode {
-	Pad     = 0, // extends image by repeating the edge of the brush
-	Repeat  = 1, // extends image by repeating the brush
-	Reflect = 2, // extends image by reflecting the brush
-};
-
-struct le_2d_linear_gradient_t {
-	uint32_t  index; // ramp index
-	glm::vec2 p0;    // start point
-	glm::vec2 p1;    // end point
-};
-
-struct le_2d_colour_stop_t {
-	float             offset; // normalized offset of the stop
-	le_2d_api::Colour colour;
-};
-
-struct Patch {
-	enum class Type : uint8_t {
-		Undefined = 0,
-		Ramp,
-		GlyphRun,
-		// Image,
-	};
-	Type type = {};
-
-	struct RampData {
-		size_t     draw_data_offset = 0; // given in count of uint32_t
-		size_t     stops_start;
-		size_t     stops_end;
-		ExtendMode extend;
-	};
-
-	struct GlyphRunData {
-		size_t index; // index of the glyph in the glyph run buffer
-	};
-
-	struct ImageData {
-		size_t draw_data_offset; // given in count of uint32_t
-		                         // Image image;
-	};
-
-	union Data {
-		RampData     as_ramp;
-		GlyphRunData as_glyph_run;
-		ImageData    as_image;
-	} data = {};
-};
-
-struct Resources {
-	// Fill this in once we want to make more advanced rendering available.
-	// this is for gradients, and images, and glyph runs.
-	std::vector<le_2d_colour_stop_t> colour_stops;
-	std::vector<Patch>               patches;
-};
-
 // ----------------------------------------------------------------------
-
-struct DrawBeginClip {
-
-	uint32_t blend_mode;
-	union {
-		float    as_float;
-		uint32_t as_uint32;
-	} alpha;
-
-	DrawBeginClip( BlendMode const& b, float a )
-	    : blend_mode( uint32_t( b.mix ) << 8 | uint32_t( b.compose ) )
-	    , alpha( a ) {
-	    };
-};
-
-// ----------------------------------------------------------------------
-struct path_encoder_o {
-
-	enum PathState : uint32_t {
-		eStart = 0,
-		eMoveTo,
-		eNonemptySubpath,
-	};
-
-	glm::vec2 first_point             = {};
-	glm::vec2 first_start_tangent_end = {};
-	PathState state                   = eStart;
-
-	uint32_t n_encoded_segments = 0;
-
-	bool is_fill = false;
-};
-
-// ----------------------------------------------------------------------
-
-struct le_2d_encoder_o : NoCopy, NoMove {
-
-	path_encoder_o path = {};
-
-	// /// The path tag stream.
-	std::vector<PathTag> path_tags;
-
-	// /// The path data stream.
-	// /// Stores all coordinates on paths.
-	std::vector<float> path_data;
-
-	// /// The draw tag stream.
-	std::vector<DrawTag> draw_tags;
-
-	// /// The draw data stream.
-	std::vector<uint32_t> draw_data; // TODO: (tig) we should use float here, so that it becomes easier to debug -- we can still do comparisons by bitfields
-
-	// /// The transform stream.
-	std::vector<Transform2D> transforms;
-
-	// /// The style stream
-	std::vector<Style> styles;
-
-	// /// Late bound resource data.
-	Resources resources; // NOTE(tig): this is not used for now as we only do full-colour paths and shapes
-
-	// /// Number of encoded paths.
-	uint32_t n_paths;
-
-	// /// Number of encoded path segments.
-	uint32_t n_path_segments;
-
-	// /// Number of encoded clips/layers.
-	uint32_t n_clips;
-
-	// /// Number of unclosed clips/layers.
-	uint32_t n_open_clips;
-
-	// /// Flags that capture the current state of the encoding.
-	uint32_t flags;
-
-	static constexpr uint32_t FORCE_NEXT_TRANSFORM = 1;
-	static constexpr uint32_t FORCE_NEXT_STYLE     = 2;
-
-	// TODO (tig): we could have an atomic counter here that makes sure that there is only ever
-	// one path encoder active; we should also assert that the atomic counter is 0 when
-	// destroying the encoder_o.
-};
 
 static void encoder_reset( le_2d_encoder_o* e ) {
 	e->path = {};
@@ -455,136 +75,6 @@ static void encoder_reset( le_2d_encoder_o* e ) {
 	e->flags           = 0;
 }
 
-template <typename T>
-inline static size_t append_to_stream( T const& src, uint8_t* target ) {
-	size_t num_bytes = src.size() * sizeof( typename T::value_type );
-	memcpy( target, src.data(), num_bytes );
-	return num_bytes;
-}
-
-// NOTE: this method should not need to be exposed - we can keep this internal to the
-// 2d renderer... only the rasterizer needs to know how to encode the data. the data
-// itself does not need to be exposed.
-//
-// static rasterizer_layout_data_t encoder_encode_to_bytes( le_2d_encoder_o const* e, std::vector<uint8_t>& bytes ) {
-static bool encoder_encode_to_bytes( le_2d_encoder_o const* e, uint8_t* bytes, size_t* bytes_count, rasterizer_layout_data_t* p_layout ) {
-
-
-	assert( e );
-
-	size_t n_path_tags     = e->path_tags.size() + e->n_open_clips;
-	size_t path_tag_padded = align_up( n_path_tags, 4 * PATH_REDUCE_WG_SZ );
-
-	// size_t path_tags_size  = n_path_tags * sizeof( decltype( e->path_tags )::value_type );
-	size_t path_data_size  = e->path_data.size() * sizeof( decltype( e->path_data )::value_type );
-	size_t draw_tags_size  = e->draw_tags.size() * sizeof( decltype( e->draw_tags )::value_type );
-	size_t open_clips_size = e->n_open_clips * sizeof( decltype( e->draw_tags )::value_type );
-	size_t draw_data_size  = e->draw_data.size() * sizeof( decltype( e->draw_data )::value_type );
-	size_t transforms_size = e->transforms.size() * sizeof( decltype( e->transforms )::value_type );
-	size_t styles_size     = e->styles.size() * sizeof( decltype( e->styles )::value_type );
-
-	size_t buffer_size =
-	    path_tag_padded +
-	    path_data_size +
-	    draw_tags_size + open_clips_size +
-	    draw_data_size +
-	    transforms_size +
-	    styles_size;
-
-	if ( nullptr == bytes_count ) {
-		return false;
-	}
-
-	if ( *bytes_count < buffer_size ) {
-		*bytes_count = buffer_size;
-		return false;
-	}
-
-	if ( nullptr == bytes ) {
-		return false;
-	}
-
-	if ( nullptr == p_layout ) {
-		return 0;
-	}
-
-	size_t used_bytes = 0;
-
-	rasterizer_layout_data_t& layout = *p_layout;
-
-	layout = {
-	    .n_paths = e->n_paths,
-	    .n_clips = e->n_clips,
-	};
-
-	layout.path_tag_base = used_bytes;
-
-	used_bytes += append_to_stream( e->path_tags, bytes + used_bytes );
-
-	if ( e->n_open_clips ) {
-		// Append any open clips as pathtag::Path
-		std::vector<PathTag> tmpOpenClips( e->n_open_clips, { PathTag::PATH } );
-		used_bytes += append_to_stream( tmpOpenClips, bytes + used_bytes );
-	}
-
-	assert( align_up( used_bytes, 4 * PATH_REDUCE_WG_SZ ) == path_tag_padded );
-
-
-	// ACHTUNG
-	// at this point, we can't just jump over the bytes that are not used,
-	// because they could contain garbage data which will confuse the gpu 
-	// -- we must zero out padding.
-	memset( bytes + used_bytes, 0, path_tag_padded - used_bytes );
-
-	used_bytes = path_tag_padded;
-
-	layout.path_data_base = used_bytes / sizeof( uint32_t );
-
-	used_bytes += append_to_stream( e->path_data, bytes + used_bytes );
-
-	layout.draw_tag_base = used_bytes / sizeof( uint32_t );
-
-	{
-		layout.bin_data_start = 0;
-
-		for ( auto const& t : e->draw_tags ) {
-			layout.bin_data_start += draw_tag_get_info_size( t );
-		}
-	}
-
-	used_bytes += append_to_stream( e->draw_tags, bytes + used_bytes );
-
-	if ( e->n_open_clips ) {
-
-		// Append any open clips as pathtag::Path
-		std::vector<DrawTag> tmpOpenClips( e->n_open_clips, { DrawTag::END_CLIP } );
-
-		used_bytes += append_to_stream( tmpOpenClips, bytes + used_bytes );
-	}
-
-	// draw data stream
-
-	layout.draw_data_base = used_bytes / sizeof( uint32_t );
-	used_bytes += append_to_stream( e->draw_data, bytes + used_bytes );
-
-	layout.transform_base = used_bytes / sizeof( uint32_t );
-
-	used_bytes += append_to_stream( e->transforms, bytes + used_bytes );
-
-	layout.style_base = used_bytes / sizeof( uint32_t );
-
-	used_bytes += append_to_stream( e->styles, bytes + used_bytes );
-
-	layout.n_drawobj = layout.n_paths;
-
-	assert( used_bytes == buffer_size );
-
-#undef vec_count_bytes
-#undef append_to_stream
-
-	*bytes_count = used_bytes;
-	return true;
-}
 
 // ----------------------------------------------------------------------
 
@@ -675,7 +165,7 @@ static void encoder_encode_linear_gradient( le_2d_encoder_o* e, le_2d_linear_gra
 
 	// ----------| invariant: more than one colour stops
 
-	// add ramp
+	// This is actually a gradient
 
 	size_t offset = e->draw_data.size(); // TODO: note : granulatity of offset is uint32_t?!
 
@@ -710,6 +200,8 @@ static void encoder_encode_linear_gradient( le_2d_encoder_o* e, le_2d_linear_gra
 	};
 
 	e->resources.patches.emplace_back( std::move( p ) );
+
+	e->draw_tags.push_back( DrawTag::LINEAR_GRADIENT );
 
 	// now, we must serialise the gradient into draw data - first allocate some memory inside draw data
 	e->draw_data.resize( offset + sizeof( le_2d_linear_gradient_t ) / sizeof( uint32_t ) );
@@ -1291,7 +783,6 @@ void register_le_2d_encoder_api( void* api_ ) {
 	encoder_i.encode_end_clip     = encoder_encode_end_clip;
 
 	//
-	encoder_i.encode_to_bytes = encoder_encode_to_bytes;
 	encoder_i.append_into_encoder = encoder_append_into_encoder;
 
 	//
